+++ module-api/src/main/kotlin/almumol/ossori/api/exception/GlobalExceptionHandler.kt
@@ -1,5 +1,6 @@
-package almumol.ossori.core.domain.project.global.exception
+package almumol.ossori.api.exception
 
+import almumol.ossori.core.domain.project.exception.BadRequestException
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.RestControllerAdvice
