+++ module-core/module-domain/src/main/kotlin/almumol/ossori/core/domain/project/exception/BadRequestException.kt
@@ -1,3 +1,3 @@
-package almumol.ossori.core.domain.project.global.exception
+package almumol.ossori.core.domain.project.exception
 
 data class BadRequestException(override val message: String) : RuntimeException()
